бот, исправления ошибок

+++ scripts/WarriorAbility/WarriorAbility.gml
@@ -1,14 +1,17 @@
 // Ресурсы скриптов были изменены для версии 2.3.0, подробности см. по адресу
 // https://help.yoyogames.com/hc/en-us/articles/360005277377
-function WarriorAbility(_using_figure=undefined, _using_cell=undefined) : FigureAbilityAction() constructor{
+function WarriorAbility(_using_figure=undefined, _using_cell=undefined, _skip_ui=false) : FigureAbilityAction() constructor{
+	skip_ui = _skip_ui;
 	using_figure = _using_figure;
 	using_cell = _using_cell;
 	target_figure = undefined;
 	target_cell = undefined;
 	
 	execute = function() {
-		using_figure.drop();
-		target_figure.drop();
+		if target_figure != undefined and target_figure.owner != using_figure.owner {
+			using_figure.drop();
+			target_figure.drop();
+		}
 	}
 	
 	draw = function() {
@@ -18,14 +21,18 @@ function WarriorAbility(_using_figure=undefined, _using_cell=undefined) : Figure
 	}
 	
 	set_target = function(_new_target, _new_cell) {
-		if target_cell != undefined {
+		if !skip_ui and target_cell != undefined {
 			target_cell.remove_figure_status(FigureStatusList.status(FIGURE_STATUS_LIST.will_be_dropped));
 		}
-		using_cell.add_figure_status(FigureStatusList.status(FIGURE_STATUS_LIST.will_be_dropped));
+		if !skip_ui {
+			using_cell.add_figure_status(FigureStatusList.status(FIGURE_STATUS_LIST.will_be_dropped));
+		}
 		target_figure = _new_target;
 		target_cell = _new_cell
-		target_cell.add_figure_status(FigureStatusList.status(FIGURE_STATUS_LIST.will_be_dropped));
-		O_BoardDraw.unblock_end_button();
+		if !skip_ui {
+			target_cell.add_figure_status(FigureStatusList.status(FIGURE_STATUS_LIST.will_be_dropped));
+			O_BoardDraw.unblock_end_button();
+		}
 	}
 	
 	check_ability_targets = function(a, b) {
@@ -34,7 +41,7 @@ function WarriorAbility(_using_figure=undefined, _using_cell=undefined) : Figure
 			for (m = -1; m <= 1; m++) {
 				cell = Game.field.get_cell(using_cell.xcord + i, using_cell.ycord + m)
 				if cell != undefined {
-					if cell.is_filled() and cell.filled_figure != using_cell.filled_figure and !cell.filled_figure.state.is_conquesting {
+					if cell.is_filled() and cell.filled_figure != using_cell.filled_figure and cell.filled_figure.owner != using_figure.owner and !cell.filled_figure.state.is_conquesting {
 						cell.marked = 1;
 						found_cell = 1;
 					}
@@ -46,6 +53,11 @@ function WarriorAbility(_using_figure=undefined, _using_cell=undefined) : Figure
 	
 	
 	back = function() {
+		if skip_ui {
+			target_figure = undefined;
+			target_cell = undefined;
+			return;
+		}
 		if target_figure != undefined {
 			global.cell_click_callback.set_draw_marks(1);
 			target_cell.remove_figure_status(FigureStatusList.status(FIGURE_STATUS_LIST.will_be_dropped));
